Deduplicate layer manifest and hash logic
Extract shared layer manifest and dependency-hash helpers into a single script module and have both build_layer.py and hash_layer.py consume it. Keep the existing script entrypoints intact while making the shared contract explicit and testable so warm-path and rebuild-path behavior cannot drift apart.

Constraint: Existing dependency hash output and deployment-type behavior must remain stable for current agents
Rejected: Leave duplicate helpers in both scripts | preserves semantic drift risk between the warm path and rebuild path
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Any future changes to dependency manifest parsing or hashing must land in the shared helper module and keep both scripts on the same contract
Tested: uv run pytest tests/unit/test_build_layer.py tests/unit/test_hash_layer.py
Tested: make preflight-session
Tested: make pre-validate-session
