+++ scripts/build_layer.py
@@ -23,12 +23,11 @@ ADRs: ADR-006
 from __future__ import annotations
 
 import argparse
-import hashlib
 import logging
 import os
 import shutil
 import subprocess
-import tomllib
+import sys
 import zipfile
 from pathlib import Path
 
@@ -41,8 +40,18 @@ logging.basicConfig(
 )
 
 REPO_ROOT = Path(__file__).resolve().parents[1]
+SCRIPT_DIR = Path(__file__).resolve().parent
+if str(SCRIPT_DIR) not in sys.path:
+    sys.path.insert(0, str(SCRIPT_DIR))
+
+import layer_manifest  # noqa: E402
+
+compute_dependency_hash = layer_manifest.compute_dependency_hash
+read_agent_deps = layer_manifest.read_agent_deps
+read_agent_lockfile = layer_manifest.read_agent_lockfile
+read_deployment_type = layer_manifest.read_deployment_type
+
 BUILD_DIR = REPO_ROOT / ".build"
-HASH_LENGTH = 16
 PYTHON_PLATFORM = "aarch64-manylinux2014"
 PYTHON_VERSION = "3.12"
 ARM64_TOKENS = ("aarch64", "arm64")
@@ -79,59 +88,6 @@ def parse_args(argv: list[str] | None = None) -> argparse.Namespace:
         help="Target environment",
     )
     return parser.parse_args(argv)
-
-
-def read_agent_deps(agent_name: str) -> list[str]:
-    """Read [project.dependencies] from agents/{agent_name}/pyproject.toml."""
-    toml_path = REPO_ROOT / "agents" / agent_name / "pyproject.toml"
-    if not toml_path.exists():
-        raise FileNotFoundError(f"pyproject.toml not found: {toml_path}")
-
-    with toml_path.open("rb") as fh:
-        data = tomllib.load(fh)
-
-    deps = data.get("project", {}).get("dependencies", [])
-    if not isinstance(deps, list):
-        raise ValueError(f"[project.dependencies] must be a list in {toml_path}")
-    return [str(dep) for dep in deps]
-
-
-def read_deployment_type(agent_name: str) -> str:
-    """Read deployment.type from agents/{agent_name}/pyproject.toml."""
-    toml_path = REPO_ROOT / "agents" / agent_name / "pyproject.toml"
-    if not toml_path.exists():
-        raise FileNotFoundError(f"pyproject.toml not found: {toml_path}")
-
-    with toml_path.open("rb") as fh:
-        data = tomllib.load(fh)
-
-    deployment = data.get("tool", {}).get("agentcore", {}).get("deployment", {})
-    deployment_type = deployment.get("type", "zip")
-    if not isinstance(deployment_type, str):
-        raise ValueError(f"[tool.agentcore.deployment.type] must be a string in {toml_path}")
-    return deployment_type
-
-
-def read_agent_lockfile(agent_name: str) -> str | None:
-    """Read uv.lock from agents/{agent_name}/uv.lock if it exists.
-
-    Returns the file content as a string, or None if the lockfile is absent.
-    """
-    lock_path = REPO_ROOT / "agents" / agent_name / "uv.lock"
-    if not lock_path.exists():
-        return None
-    return lock_path.read_text(encoding="utf-8")
-
-
-def compute_dependency_hash(deps: list[str], lockfile_content: str | None = None) -> str:
-    """Return canonical dependency hash used for S3 key and SSM metadata.
-
-    Includes lockfile content when present to track transitive dependency changes.
-    """
-    canonical = "\n".join(sorted(dep.strip() for dep in deps))
-    if lockfile_content is not None:
-        canonical = canonical + "\n---lockfile---\n" + lockfile_content
-    return hashlib.sha256(canonical.encode()).hexdigest()[:HASH_LENGTH]
 
 
 def build_dependencies(dependencies: list[str], target_dir: Path) -> None:
