+++ scripts/hash_layer.py
@@ -27,10 +27,9 @@ ADRs: ADR-006, ADR-008
 from __future__ import annotations
 
 import argparse
-import hashlib
 import logging
 import os
-import tomllib
+import sys
 from pathlib import Path
 
 import boto3
@@ -43,63 +42,16 @@ logging.basicConfig(
 )
 
 REPO_ROOT = Path(__file__).resolve().parents[1]
-HASH_LENGTH = 16
+SCRIPT_DIR = Path(__file__).resolve().parent
+if str(SCRIPT_DIR) not in sys.path:
+    sys.path.insert(0, str(SCRIPT_DIR))
 
+import layer_manifest  # noqa: E402
 
-def compute_dependency_hash(deps: list[str], lockfile_content: str | None = None) -> str:
-    """Return a canonical SHA256 hash of dependencies and lockfile state.
-
-    Canonical form: each dep entry stripped of whitespace, sorted, joined by
-    newline.  When lockfile content is provided it is appended after a separator.
-    Returns the first HASH_LENGTH hex characters of SHA256.
-    Same deps in any order produce the same hash.
-    """
-    canonical = "\n".join(sorted(d.strip() for d in deps))
-    if lockfile_content is not None:
-        canonical = canonical + "\n---lockfile---\n" + lockfile_content
-    return hashlib.sha256(canonical.encode()).hexdigest()[:HASH_LENGTH]
-
-
-def read_agent_deps(agent_name: str) -> list[str]:
-    """Read [project.dependencies] from agents/{agent_name}/pyproject.toml."""
-    toml_path = REPO_ROOT / "agents" / agent_name / "pyproject.toml"
-    if not toml_path.exists():
-        raise FileNotFoundError(f"pyproject.toml not found: {toml_path}")
-
-    with toml_path.open("rb") as fh:
-        data = tomllib.load(fh)
-
-    deps = data.get("project", {}).get("dependencies", [])
-    if not isinstance(deps, list):
-        raise ValueError(f"[project.dependencies] must be a list in {toml_path}")
-    return [str(d) for d in deps]
-
-
-def read_deployment_type(agent_name: str) -> str:
-    """Read deployment.type from agents/{agent_name}/pyproject.toml."""
-    toml_path = REPO_ROOT / "agents" / agent_name / "pyproject.toml"
-    if not toml_path.exists():
-        raise FileNotFoundError(f"pyproject.toml not found: {toml_path}")
-
-    with toml_path.open("rb") as fh:
-        data = tomllib.load(fh)
-
-    deployment = data.get("tool", {}).get("agentcore", {}).get("deployment", {})
-    deployment_type = deployment.get("type", "zip")
-    if not isinstance(deployment_type, str):
-        raise ValueError(f"[tool.agentcore.deployment.type] must be a string in {toml_path}")
-    return deployment_type
-
-
-def read_agent_lockfile(agent_name: str) -> str | None:
-    """Read uv.lock from agents/{agent_name}/uv.lock if it exists.
-
-    Returns the file content as a string, or None if the lockfile is absent.
-    """
-    lock_path = REPO_ROOT / "agents" / agent_name / "uv.lock"
-    if not lock_path.exists():
-        return None
-    return lock_path.read_text(encoding="utf-8")
+compute_dependency_hash = layer_manifest.compute_dependency_hash
+read_agent_deps = layer_manifest.read_agent_deps
+read_agent_lockfile = layer_manifest.read_agent_lockfile
+read_deployment_type = layer_manifest.read_deployment_type
 
 
 def get_ssm_hash(agent_name: str, env: str, aws_region: str) -> str | None:
